feat(metaengine): use DOUBLE for float64, surface reification failures, and fix key/value field collisions
- Change float64 column type from REAL to DOUBLE so DuckDB preserves full float64 precision; REAL would silently truncate to float32 and corrupt aggregation results
- Update layout planner docs and tests to reflect the INTEGER/DOUBLE type mapping
- Fix bug in BuildColumnarLayoutPlan where struct fields named "Key" or "Value" collided with the base table's `key`/`value` columns on case-insensitive engines (DuckDB, SQLite); collision produced DDL errors at table creation
- Add ReificationFailures counter to WorkloadStats so consumers can detect watcher values that fail to reify to the declared type V; clearly documented as a health diagnostic, not a planner input
- Expose reification failure count from workloadMeter.Stats() and assert it in the existing typesafe watcher test
- Replace generic parameter in WithColumnarLayout[ProductView]() with non-generic WithColumnarLayout(); the result type is already carried in the layout plan, so the type parameter was redundant
- Replace round-number test prices (1.0, 0.5, 1.5, 2.0) with realistic decimals (1.99, 0.75, 0.99, 2.49) so aggregation tests would fail under float32 truncation and therefore prove DOUBLE precision end-to-end
- Add TestDuckDBEngine_ColumnarDoublePrecision which round-trips pi (3.141592653589793) through a planned columnar layout and asserts the exact value is preserved; a float32-backed REAL column would truncate to 3.1415927
- Reorder Preset struct tags in cqrs-lint config to match canonical Go tag ordering (default before json)

+++ metaengine/layout_type_test.go
@@ -29,7 +29,7 @@ func TestBuildLayoutPlanFromType_InfersColumnTypes(t *testing.T) {
 	cases := map[string]string{
 		"status":   "TEXT",    // string
 		"priority": "INTEGER", // int
-		"score":    "REAL",    // float64
+		"score":    "DOUBLE",  // float64
 		"name":     "TEXT",    // string
 	}
